+++ modules/System/views/info.php
@@ -107,6 +107,7 @@
                             <tr><td>PHP SAPI</td><td class="kiss-color-muted"><?=php_sapi_name()?></td></tr>
                             <tr><td>System</td><td class="kiss-color-muted"><?=php_uname()?></td></tr>
                             <tr><td>Extensions</td><td class="kiss-color-muted"><?=implode(', ', get_loaded_extensions())?></td></tr>
+                            <tr><td>Supported image types</td><td class="kiss-color-muted"><?=implode(', ', $supportedImageTypes)?></td></tr>
                             <tr><td>Max. execution time</td><td class="kiss-color-muted"><?=ini_get('max_execution_time')?> sec.</td></tr>
                             <tr><td>Memory limit</td><td class="kiss-color-muted"><?=ini_get("memory_limit")?></td></tr>
                             <tr><td>Upload file size limit</td><td class="kiss-color-muted"><?=ini_get("upload_max_filesize")?></td></tr>